pyproject.toml
[tool.poetry]
name = "pokemon"
version = "0.1.0"
description = ""
authors = ["Ash <ashtokyo31@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.1"
django-environ = "^0.4.5"
psycopg2 = "^2.8.6"
drf-yasg = "^1.20.0"


[tool.poetry.dev-dependencies]
pudb = "^2019.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"